Server/api/requirements.txt
 Django
 djangorestframework
 markdown
 django-filter
 psycopg2>=2.7,<3.0
 pika
 pyzmq
 django-cors-headers
 